Create sr_latin.yaml
Add Serbian language support (same as Bosnian, Croatian, and Montenegro)  

This PR adds support for the Serbian language, which is interchangeable with Bosnian, Croatian, and Macedonian ( in the same way that German is used in both Germany and Austria. https://en.wikipedia.org/wiki/Serbo-Croatian )

Since this is my first PR, I would appreciate it if the maintainer could review it for any potential mistakes.